+++ packages/slonik/src/service.ts
@@ -107,7 +107,7 @@ abstract class BaseService<
     offset?: number,
     filters?: FilterInput,
     sort?: SortInput[]
-  ): Promise<readonly T[]> => {
+  ): Promise<PaginatedList<T>> => {
     const query = this.factory.getListSql(
       Math.min(limit ?? this.getLimitDefault(), this.getLimitMax()),
       offset,
@@ -115,33 +115,23 @@ abstract class BaseService<
       sort
     );
 
-    const result = await this.database.connect((connection) => {
-      return connection.any(query);
-    });
+    const [totalCount, filteredCount, data] = await Promise.all([
+      this.count(),
+      this.count(filters),
+      this.database.connect((connection) => {
+        return connection.any(query);
+      }),
+    ]);
 
-    return result as T[];
-  };
-
-  paginatedList = async (
-    limit?: number,
-    offset?: number,
-    filters?: FilterInput,
-    sort?: SortInput[]
-  ): Promise<PaginatedList<T>> => {
-    const listResult = await this.list(limit, offset, filters, sort);
-
-    const countResult = await this.count(filters);
-
-    const combinedResult = {
-      totalCount: countResult,
-      data: [...listResult],
+    return {
+      totalCount,
+      filteredCount,
+      data,
     };
-
-    return combinedResult as PaginatedList<T>;
   };
 
   count = async (filters?: FilterInput): Promise<number> => {
-    const query = this.factory.getCount(filters);
+    const query = this.factory.getCountSql(filters);
 
     const result = await this.database.connect((connection) => {
       return connection.any(query);