+++ packages/slonik/src/types/service.ts
@@ -20,19 +20,15 @@ interface Service<T, C, U> {
     offset?: number,
     filters?: FilterInput,
     sort?: SortInput[]
-  ): Promise<readonly T[]>;
-  paginatedList(
-    limit?: number,
-    offset?: number,
-    filters?: FilterInput,
-    sort?: SortInput[]
   ): Promise<PaginatedList<T>>;
   count(filters?: FilterInput): Promise<number>;
   update(id: number | string, data: U): Promise<T>;
 }
 
-type PaginatedList<T> = { totalCount: number; data: readonly T[] };
+type PaginatedList<T> = {
+  totalCount: number;
+  filteredCount: number;
+  data: readonly T[];
+};
 
-export type { PaginatedList };
-
-export type { Service };
+export type { PaginatedList, Service };