+++ packages/user/src/model/users/resolver.ts
@@ -88,7 +88,7 @@ const Query = {
   ) => {
     const service = new Service(context.config, context.database);
 
-    return await service.paginatedList(
+    return await service.list(
       arguments_.limit,
       arguments_.offset,
       arguments_.filters