+++ packages/user/src/supertokens/recipes/config/third-party-email-password/emailPasswordSignIn.ts
@@ -1,5 +1,3 @@
-import UserRoles from "supertokens-node/recipe/userroles";
-
 import UserService from "../../../../model/user-profiles/service";
 
 import type { User, UserCreateInput, UserUpdateInput } from "../../../../types";
@@ -10,7 +8,7 @@ import type { RecipeInterface } from "supertokens-node/recipe/thirdpartyemailpas
 const emailPasswordSignIn = (
   originalImplementation: RecipeInterface,
   fastify: FastifyInstance
-): typeof originalImplementation.emailPasswordSignIn => {
+): RecipeInterface["emailPasswordSignIn"] => {
   const { config, slonik } = fastify;
 
   return async (input) => {