+++ packages/user/src/supertokens/recipes/config/third-party-email-password/emailPasswordSignUpPost.ts
@@ -1,5 +1,3 @@
-import UserRoles from "supertokens-node/recipe/userroles";
-
 import type { User } from "../../../../types";
 import type { FastifyInstance } from "fastify";
 import type { APIInterface } from "supertokens-node/recipe/thirdpartyemailpassword/types";
@@ -8,7 +6,7 @@ const emailPasswordSignUpPOST = (
   originalImplementation: APIInterface,
   /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
   fastify: FastifyInstance
-): typeof originalImplementation.emailPasswordSignUpPOST => {
+): APIInterface["emailPasswordSignUpPOST"] => {
   return async (input) => {
     if (originalImplementation.emailPasswordSignUpPOST === undefined) {
       throw new Error("Should never come here");