+++ packages/user/src/supertokens/recipes/config/third-party-email-password/sendEmail.ts
@@ -1,31 +1,40 @@
 import ThirdPartyEmailPassword from "supertokens-node/recipe/thirdpartyemailpassword";
 
+import getOrigin from "../../../utils/getOrigin";
 import mailer from "../../../utils/sendEmail";
 
-import type { FastifyInstance } from "fastify";
+import type { FastifyInstance, FastifyRequest } from "fastify";
+import type { EmailDeliveryInterface } from "supertokens-node/lib/build/ingredients/emaildelivery/types";
+import type { TypeEmailPasswordPasswordResetEmailDeliveryInput } from "supertokens-node/lib/build/recipe/emailpassword/types";
 
 const sendEmail = (
+  originalImplementation: EmailDeliveryInterface<TypeEmailPasswordPasswordResetEmailDeliveryInput>,
   fastify: FastifyInstance
 ): typeof ThirdPartyEmailPassword.sendEmail => {
-  const { config } = fastify;
-
-  const websiteDomain = config.appOrigin[0] as string;
+  const websiteDomain = fastify.config.appOrigin[0] as string;
   const resetPasswordPath = "/reset-password";
 
   return async (input) => {
+    const request: FastifyRequest = input.userContext._default.request.request;
+
+    const url =
+      request.headers.referer || request.headers.origin || request.hostname;
+
+    const origin = getOrigin(url) || websiteDomain;
+
+    const passwordResetLink = input.passwordResetLink.replace(
+      websiteDomain + "/auth/reset-password",
+      origin +
+        (fastify.config.user.supertokens.resetPasswordPath || resetPasswordPath)
+    );
+
     await mailer({
       fastify,
       subject: "Reset Password",
       templateName: "reset-password",
       to: input.user.email,
       templateData: {
-        passwordResetLink: input.passwordResetLink.replace(
-          websiteDomain + "/auth/reset-password",
-          websiteDomain +
-            (config.user.supertokens.resetPasswordPath
-              ? (config.user.supertokens.resetPasswordPath as string)
-              : resetPasswordPath)
-        ),
+        passwordResetLink,
       },
     });
   };