+++ packages/user/src/supertokens/types.ts
@@ -1,18 +1,44 @@
 import ThirdPartyEmailPassword from "supertokens-node/recipe/thirdpartyemailpassword";
 
 import type { FastifyInstance } from "fastify";
+import type { EmailDeliveryInterface } from "supertokens-node/lib/build/ingredients/emaildelivery/types";
+import type { TypeEmailPasswordPasswordResetEmailDeliveryInput } from "supertokens-node/lib/build/recipe/emailpassword/types";
 import type { TypeInput as SessionRecipeConfig } from "supertokens-node/recipe/session/types";
-import type { TypeInput as ThirdPartyEmailPasswordRecipeConfig } from "supertokens-node/recipe/thirdpartyemailpassword/types";
+import type {
+  TypeInput as ThirdPartyEmailPasswordRecipeConfig,
+  APIInterface,
+  RecipeInterface,
+  TypeInputSignUp,
+} from "supertokens-node/recipe/thirdpartyemailpassword/types";
 import type { TypeInput as UserRolesRecipeConfig } from "supertokens-node/recipe/userroles/types";
 
 const { Apple, Facebook, Github, Google } = ThirdPartyEmailPassword;
 
+type APIInterfaceWrapper = {
+  [key in keyof APIInterface]?: (
+    originalImplementation: APIInterface,
+    fastify: FastifyInstance
+  ) => APIInterface[key];
+};
+
+type SendEmailWrapper = (
+  originalImplementation: EmailDeliveryInterface<TypeEmailPasswordPasswordResetEmailDeliveryInput>,
+  fastify: FastifyInstance
+) => typeof ThirdPartyEmailPassword.sendEmail;
+
+type RecipeInterfaceWrapper = {
+  [key in keyof RecipeInterface]?: (
+    originalImplementation: RecipeInterface,
+    fastify: FastifyInstance
+  ) => RecipeInterface[key];
+};
+
 interface SupertokensRecipes {
   session?: (fastify: FastifyInstance) => SessionRecipeConfig;
   userRoles?: (fastify: FastifyInstance) => UserRolesRecipeConfig;
-  thirdPartyEmailPassword?: (
-    fastify: FastifyInstance
-  ) => ThirdPartyEmailPasswordRecipeConfig;
+  thirdPartyEmailPassword?:
+    | ThirdPartyEmailPasswordRecipe
+    | ((fastify: FastifyInstance) => ThirdPartyEmailPasswordRecipeConfig);
 }
 
 interface SupertokensThirdPartyProvider {
@@ -20,6 +46,15 @@ interface SupertokensThirdPartyProvider {
   facebook?: Parameters<typeof Facebook>[0];
   github?: Parameters<typeof Github>[0];
   google?: Parameters<typeof Google>[0];
+}
+
+interface ThirdPartyEmailPasswordRecipe {
+  override?: {
+    apis?: APIInterfaceWrapper;
+    functions?: RecipeInterfaceWrapper;
+  };
+  sendEmail?: SendEmailWrapper;
+  signUpFeature?: TypeInputSignUp;
 }
 
 interface SupertokensConfig {
@@ -30,4 +65,11 @@ interface SupertokensConfig {
   sendUserAlreadyExistsWarning?: boolean;
 }
 
-export type { SupertokensConfig, SupertokensRecipes };
+export type {
+  APIInterfaceWrapper,
+  RecipeInterfaceWrapper,
+  SendEmailWrapper,
+  SupertokensConfig,
+  SupertokensRecipes,
+  ThirdPartyEmailPasswordRecipe,
+};